Move BeatsPerMinute and replace_false_apostrophes_and_quotation_marks()
to song_txt.auxiliaries to avoid cyclic imports.
